Added viewModel for organization position dialog

+++ app/src/main/java/com/elka/heofficeclub/service/model/OrganizationPosition.kt
@@ -3,5 +3,5 @@ package com.elka.heofficeclub.service.model
 data class OrganizationPosition(
   var id: String = "",
   var name: String = "",
-  var salary: Int = 0,
+  var salary: Double = 0.0,
 )
